feat: enhance Discord role management; add support for role attributes and improve permission resolution

+++ plugins/plugin-discord-bot/src/main/java/irden/space/proxy/plugin/discord/model/DiscordRankRoles.java
@@ -2,7 +2,7 @@ package irden.space.proxy.plugin.discord.model;
 
 public record DiscordRankRoles(
         String name,
-        int roleId
+        long roleId
 ) {
 
 }
